+++ src/pages/EventCreate.tsx
@@ -20,8 +20,8 @@ import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import dayjs, { Dayjs } from "dayjs";
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { apiCreateEvent, apiGetSports } from "./api/SpentApiManager";
-import { Sport } from "./types/types";
+import { apiCreateEvent, apiGetSports } from "../api/SpentApiManager";
+import { Sport } from "../types/types";
 
 export default function EventCreate() {
   const navigate = useNavigate();
